Revert "feat(ui): split the sidebar into tools and projects (#465)"
This reverts commit 9b1c20938e25d5a094792d01d3a191c1bf3d2541.
