+++ package.nls.json
@@ -17,6 +17,5 @@
   "configuration.playwright.env": "Environment variables to pass to Playwright Test.",
   "configuration.playwright.reuseBrowser": "Show & reuse browser between tests.",
   "configuration.playwright.showTrace": "Show Trace Viewer.",
-  "views.test.pw.extension.settingsView": "Playwright",
-  "views.test.pw.extension.projectsView": "Playwright Projects"
+  "views.test.pw.extension.settingsView": "Playwright"
 }
